Implement first bound function
Add prim branch & bound

Remove session.xml

Various improvements

Prim bound fixes

Prim bound fixes

Remove print

Debug info

Fix prim

Put back tsp3

+++ src/main/java/fr/blazanome/routeplanner/algorithm/tsp/TSP1.java
@@ -9,7 +9,7 @@ import java.util.stream.StreamSupport;
 
 public class TSP1 extends TemplateTSP {
 	@Override
-	protected int bound(Integer currentVertex, Collection<Integer> unvisited) {
+	protected double bound(Integer currentVertex, Collection<Integer> unvisited) {
 		return 0;
 	}
